backend/requirements.txt
flask
parser
mysql-connector
